requirements.txt
protobuf == 3.20
numpy==1.21.6
openai==0.27.8
Pillow==9.5.0
streamlit==1.23.0
tensorflow==2.10.0

